Add files via upload
A sequence of steps are followed to play the music using the remote control. If the user is not authenticated (through the finger print touch sensor), they will not be able to see the songs list. Also, if the sequence of buttons are not pressed correctly, users will not be able to play the songs.